Revert "Database queries tracing with datasource-proxy. (#1095)"
This reverts commit 972a01bff451da0aeb16a4e016ebe0acf44a8c55.
