--- sentry-p6spy/build.gradle.kts
@@ -1,66 +1,0 @@
-import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
-
-plugins {
-    `java-library`
-    kotlin("jvm")
-    jacoco
-    id(Config.QualityPlugins.errorProne)
-    id(Config.QualityPlugins.gradleVersions)
-}
-
-configure<JavaPluginConvention> {
-    sourceCompatibility = JavaVersion.VERSION_1_8
-    targetCompatibility = JavaVersion.VERSION_1_8
-}
-
-tasks.withType<KotlinCompile>().configureEach {
-    kotlinOptions.jvmTarget = JavaVersion.VERSION_1_8.toString()
-}
-
-dependencies {
-    api(project(":sentry"))
-    api(Config.Libs.p6spy)
-
-    compileOnly(Config.CompileOnly.nopen)
-    errorprone(Config.CompileOnly.nopenChecker)
-    errorprone(Config.CompileOnly.errorprone)
-    errorproneJavac(Config.CompileOnly.errorProneJavac8)
-    compileOnly(Config.CompileOnly.jetbrainsAnnotations)
-
-    // tests
-    testImplementation(project(":sentry-test-support"))
-    testImplementation(kotlin(Config.kotlinStdLib))
-    testImplementation(Config.TestLibs.kotlinTestJunit)
-    testImplementation(Config.TestLibs.mockitoKotlin)
-    testImplementation(Config.TestLibs.awaitility)
-    testImplementation(Config.TestLibs.hsqldb)
-}
-
-configure<SourceSetContainer> {
-    test {
-        java.srcDir("src/test/java")
-    }
-}
-
-jacoco {
-    toolVersion = Config.QualityPlugins.Jacoco.version
-}
-
-tasks.jacocoTestReport {
-    reports {
-        xml.isEnabled = true
-        html.isEnabled = false
-    }
-}
-
-tasks {
-    jacocoTestCoverageVerification {
-        violationRules {
-            rule { limit { minimum = Config.QualityPlugins.Jacoco.minimumCoverage } }
-        }
-    }
-    check {
-        dependsOn(jacocoTestCoverageVerification)
-        dependsOn(jacocoTestReport)
-    }
-}
