+++ sentry-samples/sentry-samples-spring-boot/api/sentry-samples-spring-boot.api
@@ -20,7 +20,7 @@ public class io/sentry/samples/spring/boot/PersonController {
 }
 
 public class io/sentry/samples/spring/boot/PersonService {
-	public fun <init> (Lorg/springframework/jdbc/core/JdbcTemplate;)V
+	public fun <init> ()V
 }
 
 public class io/sentry/samples/spring/boot/SecurityConfiguration : org/springframework/security/config/annotation/web/configuration/WebSecurityConfigurerAdapter {
