+++ sentry-samples/sentry-samples-spring-boot/src/main/resources/application.properties
@@ -8,9 +8,5 @@ sentry.logging.minimum-event-level=info
 sentry.logging.minimum-breadcrumb-level=debug
 # Performance configuration
 sentry.traces-sample-rate=1.0
-# Database configuration
-spring.datasource.url=jdbc:hsqldb:mem:testdb
-spring.datasource.username=sa
-spring.datasource.password=
 sentry.debug=true
 in-app-includes="io.sentry.samples"
