--- sentry-samples/sentry-samples-spring-boot/src/main/resources/schema.sql
@@ -1,5 +1,0 @@
-CREATE TABLE person (
-   id INTEGER IDENTITY PRIMARY KEY,
-   firstName VARCHAR(50) NOT NULL,
-   lastName VARCHAR(50) NOT NULL
-);
