+++ sentry-spring-boot-starter/src/main/java/io/sentry/spring/boot/SentryAutoConfiguration.java
@@ -26,6 +26,8 @@ import org.aspectj.lang.ProceedingJoinPoint;
 import org.jetbrains.annotations.NotNull;
 import org.springframework.beans.factory.ObjectProvider;
 import org.springframework.boot.autoconfigure.AutoConfigureBefore;
+import org.springframework.boot.autoconfigure.condition.AnyNestedCondition;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
@@ -208,5 +210,24 @@ public class SentryAutoConfiguration {
 
       return sdkVersion;
     }
+  }
+
+  static final class SentryTracingCondition extends AnyNestedCondition {
+
+    public SentryTracingCondition() {
+      super(ConfigurationPhase.REGISTER_BEAN);
+    }
+
+    @ConditionalOnProperty(name = "sentry.enable-tracing", havingValue = "true")
+    @SuppressWarnings("UnusedNestedClass")
+    private static class SentryTracingEnabled {}
+
+    @ConditionalOnProperty(name = "sentry.traces-sample-rate")
+    @SuppressWarnings("UnusedNestedClass")
+    private static class SentryTracesSampleRateCondition {}
+
+    @ConditionalOnBean(SentryOptions.TracesSamplerCallback.class)
+    @SuppressWarnings("UnusedNestedClass")
+    private static class SentryTracesSamplerBeanCondition {}
   }
 }
